changed parameter type of convertMilesToCash from double to int

+++ src/main/java/RewardValue.java
@@ -18,7 +18,7 @@ public class RewardValue {
         return milesValue;
     }
 
-    private double convertMilesToCash(double milesValue) {
+    private double convertMilesToCash(int milesValue) {
         return milesValue * conversionRate;
     }
